organize stacks directories
