+++ assets/src/components/stacks/prs/StackPrs.tsx
@@ -1,0 +1,8 @@
+import { Card } from '@pluralsh/design-system'
+import { useParams } from 'react-router-dom'
+
+export function StackPrs() {
+  const { stackId = '' } = useParams()
+
+  return <Card>Stack Prs: {stackId}</Card>
+}
